Implemented Helper Functions:
parse data - parse data to the Item class where it is managed.
validateItem - validate so a bad item isn't added to the list
clear form function - will clear all input in the form.
custom alert function to be used throughout the class - eventually make
it so it has a picture on it e.g a tick for a successfully added item.